requirements.txt
certifi==2019.3.9
chardet==3.0.4
decorator==4.4.2
distlib==0.3.0
Django==2.2.13
django-admin-tools==0.9.0
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
gunicorn==20.0.4
json5==0.9.4
jsonschema==3.2.0
pipenv==2018.11.26
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.11.3
psycopg2==2.8.4
Pillow==5.2.0
python-dateutil==2.8.1
pytz==2018.5
requests==2.21.0
rsa==4.0
scipy==1.4.1
Send2Trash==1.5.0
six==1.13.0
sqlparse==0.2.4
urllib3==1.24.2
virtualenv==20.0.18
virtualenv-clone==0.5.4